підключення моделі для даних про сон

+++ Myapi/Controlers/UserPageController.cs
@@ -16,11 +16,16 @@ public class UserPageController : ControllerBase
 
     public async Task<IActionResult> GetUser(string id)
     {
-        var user = await _userService.GetByIdAsync(id);
-        if (user == null)
-            return NotFound();
+         var userId = HttpContext.Session.GetString("UserEmail");
 
-        return Ok(user);
+          if (userId == null)
+        return Unauthorized(new { message = "Not logged in" });
+
+        var user = await _userService.GetByEmailAsync(userId);
+        if (user == null)
+        return NotFound();
+
+    return Ok(user);
     }
     [HttpPut("setPhoto/{id}")]
     public async Task<IActionResult> SetPhoto(string id, IFormFile photo)
@@ -40,15 +45,19 @@ public class UserPageController : ControllerBase
         {
             var oldFilePath = Path.Combine(folderPath, user.Photo);
             if (System.IO.File.Exists(oldFilePath))
-            {
-                try
+            { 
+                    try
                 {
-                    System.IO.File.Delete(oldFilePath);
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine($"Помилка при видаленні файлу: {ex.Message}");
-                }
+                        // if (Path.GetFileName(oldFilePath) != "avatar.png"){
+                        //       System.IO.File.Delete(oldFilePath);
+                          if (user.Photo != "avatar.png"){
+                              System.IO.File.Delete(oldFilePath);
+                    }
+                    }
+                    catch (Exception ex)
+                    {
+                        Console.WriteLine($"Помилка при видаленні файлу: {ex.Message}");
+                    }
             }
         }
         using (var stream = new FileStream(filePath, FileMode.Create))
@@ -57,6 +66,28 @@ public class UserPageController : ControllerBase
         }
         await _userService.UpdatePhotoAsync(id, fileName);
 
-        return Ok(new { message = "Фото оновлено успішно", photo = fileName });
+        // return Ok(new { message = "Фото оновлено успішно", photo = fileName });
+        return Ok(new { message = "Фото оновлено успішно" });
+    }
+    [HttpDelete("delete/{id}")]
+    public async Task<IActionResult> delete(string id)
+    {
+        var user = await _userService.GetByIdAsync(id);
+        if (user == null)
+            return NotFound(new { message = "Користувача не знайдено" });
+        await _userService.DeleteAsync(id);
+        return Ok("Акаунт видалено");
+    }
+    [HttpPost("logout")]
+    public IActionResult Logout()
+    {
+        HttpContext.Session.Clear();
+        return Ok(new { message = "Ви вийшли з акаунту" });
+    }
+    [HttpGet("check-session")]
+    public IActionResult CheckSession()
+    {
+    var user = HttpContext.Session.GetString("UserEmail");
+    return Ok(new { UserEmail = user });
     }
 }
